if nopg, guess for nodes without headpage based on the headpage setting
If using default settings, `:start:` is always first, so behaviour is
unchanged. It uses the first enabled option if multiple options are
enabled. Improves #106

Edge case: if all headpage options are disabled and no custom headpage
name is given, it resulted in an empty string value, which was converted
in ns1: which is in general interpreted as ns1:<startpage>. Custom
headpage should be not blank.