Add data-wiki-id attr to Fancytree (Treenew)
DokuWiki adds this attribute to "a" interwikis automatically, it points to the namespace/pageID, e.g. "doku.php?id=3cx:add_phones" the value of data-wiki-id attr is "3cx:add_phones".
On Read the Dokus template, this attribute is used on sidebar to determine the "Next" and "Previous" buttons location.