Open links on a new tab when Ctrl (Win) / Command (macOS) key is pressed
When you're editing a page (&do=edit) and attempt to open another page through the sidebar pressing Ctrl (Windows) or Command (macOS) key, the plug-in attempts to load the clicked page on the current tab, triggering the alert "Changes not saved". This commit adds support for Ctrl (Win)/Command (macOS) key, so if those keys are pressed, instead of opening the link on the current tab, it opens the clicked link on a new tab.